+++ test/clojure/lazytest/experimental/interfaces/xunit_test.clj
@@ -1,6 +1,12 @@
-(ns lazytest.experimental.interfaces.xunit-test 
+(ns lazytest.experimental.interfaces.xunit-test
   (:require
-    [lazytest.experimental.interfaces.xunit :refer [defsuite suite test-case assert!]]))
+   [lazytest.core :refer [defdescribe expect it]]
+   [lazytest.extensions.matcher-combinators :refer [match?]]))
+
+(in-ns 'xunit-temp)
+(clojure.core/refer 'clojure.core)
+(clojure.core/require
+ '[lazytest.experimental.interfaces.xunit :refer [defsuite suite test-case assert!]])
 
 (defsuite defsuite-test
   (suite "defsuite works"
@@ -8,3 +14,26 @@
   (suite "suite works"
     (test-case "test-case works"
       (assert! true))))
+
+(in-ns 'lazytest.experimental.interfaces.xunit-test)
+
+#_{:clj-kondo/ignore [:unresolved-namespace]}
+(let [suite (xunit-temp/defsuite-test)]
+  (defn existing-tests [] suite))
+
+(remove-ns 'xunit-temp)
+
+(defdescribe xunit-tests
+  (it "has the right shape"
+    (expect
+      (match?
+       {:type :lazytest/var
+        :children
+        [{:type :lazytest/suite
+          :doc "defsuite works"}
+         {:type :lazytest/suite
+          :doc "suite works"
+          :children
+          [{:type :lazytest/test-case
+            :doc "test-case works"}]}]}
+       (existing-tests)))))